+++ src/canisters/frontend/ic-certified-assets/src/state_machine.rs
@@ -1,5 +1,9 @@
 //! This module contains a pure implementation of the certified assets state machine.
 
+mod v2;
+
+pub use v2::StableState as StableStateV2;
+
 // NB. This module should not depend on ic_cdk, it contains only pure state transition functions.
 // All the environment (time, certificates, etc.) is passed to the state transition functions
 // as formal arguments.  This approach makes it very easy to test the state machine.
@@ -69,14 +73,6 @@ const DEFAULT_MAX_COMPUTE_EVIDENCE_ITERATIONS: u16 = 20;
 
 type Timestamp = Int;
 
-fn timestamp_to_u64(timestamp: Timestamp) -> u64 {
-    timestamp.0.to_u64().expect("timestamp overflow")
-}
-
-fn batch_id_to_u64(batch_id: BatchId) -> u64 {
-    batch_id.0.to_u64().expect("batch id overflow")
-}
-
 #[derive(Default, Clone, Debug, CandidType, Deserialize)]
 pub struct AssetEncoding {
     pub modified: Timestamp,
@@ -88,46 +84,6 @@ pub struct AssetEncoding {
     pub sha256: [u8; 32],
     pub certificate_expression: Option<CertificateExpression>,
     pub response_hashes: Option<HashMap<u16, [u8; 32]>>,
-}
-
-/// Same as [AssetEncoding] but serializable with cbor
-#[derive(Default, Clone, Debug, Deserialize, Serialize)]
-pub struct StableAssetEncodingV2 {
-    pub modified: u64,
-    pub content_chunks: Vec<RcBytes>,
-    pub total_length: usize,
-    pub certified: bool,
-    pub sha256: [u8; 32],
-    pub certificate_expression: Option<CertificateExpression>,
-    pub response_hashes: Option<HashMap<u16, [u8; 32]>>,
-}
-
-impl From<AssetEncoding> for StableAssetEncodingV2 {
-    fn from(asset_encoding: AssetEncoding) -> Self {
-        Self {
-            modified: timestamp_to_u64(asset_encoding.modified),
-            content_chunks: asset_encoding.content_chunks,
-            total_length: asset_encoding.total_length,
-            certified: asset_encoding.certified,
-            sha256: asset_encoding.sha256,
-            certificate_expression: asset_encoding.certificate_expression,
-            response_hashes: asset_encoding.response_hashes,
-        }
-    }
-}
-
-impl From<StableAssetEncodingV2> for AssetEncoding {
-    fn from(stable_asset_encoding: StableAssetEncodingV2) -> Self {
-        Self {
-            modified: Timestamp::from(stable_asset_encoding.modified),
-            content_chunks: stable_asset_encoding.content_chunks,
-            total_length: stable_asset_encoding.total_length,
-            certified: stable_asset_encoding.certified,
-            sha256: stable_asset_encoding.sha256,
-            certificate_expression: stable_asset_encoding.certificate_expression,
-            response_hashes: stable_asset_encoding.response_hashes,
-        }
-    }
 }
 
 impl AssetEncoding {
@@ -208,51 +164,6 @@ pub struct Asset {
     pub allow_raw_access: Option<bool>,
 }
 
-/// Same as [Asset] but serializable with cbor
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct StableAssetV2 {
-    pub content_type: String,
-    pub encodings: HashMap<String, StableAssetEncodingV2>,
-    pub max_age: Option<u64>,
-    pub headers: Option<HashMap<String, String>>,
-    pub is_aliased: Option<bool>,
-    pub allow_raw_access: Option<bool>,
-}
-
-impl From<Asset> for StableAssetV2 {
-    fn from(asset: Asset) -> Self {
-        Self {
-            content_type: asset.content_type,
-            encodings: asset
-                .encodings
-                .into_iter()
-                .map(|(k, v)| (k, v.into()))
-                .collect(),
-            max_age: asset.max_age,
-            headers: asset.headers,
-            is_aliased: asset.is_aliased,
-            allow_raw_access: asset.allow_raw_access,
-        }
-    }
-}
-
-impl From<StableAssetV2> for Asset {
-    fn from(stable_asset: StableAssetV2) -> Self {
-        Self {
-            content_type: stable_asset.content_type,
-            encodings: stable_asset
-                .encodings
-                .into_iter()
-                .map(|(k, v)| (k, v.into()))
-                .collect(),
-            max_age: stable_asset.max_age,
-            headers: stable_asset.headers,
-            is_aliased: stable_asset.is_aliased,
-            allow_raw_access: stable_asset.allow_raw_access,
-        }
-    }
-}
-
 #[derive(Clone, Debug, CandidType, Deserialize)]
 pub struct EncodedAsset {
     pub content: RcBytes,
@@ -302,34 +213,6 @@ pub struct Configuration {
     pub max_bytes: Option<u64>,
 }
 
-/// Same as [Configuration] but serializable with cbor
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct StableConfigurationV2 {
-    pub max_batches: Option<u64>,
-    pub max_chunks: Option<u64>,
-    pub max_bytes: Option<u64>,
-}
-
-impl From<Configuration> for StableConfigurationV2 {
-    fn from(configuration: Configuration) -> Self {
-        Self {
-            max_batches: configuration.max_batches,
-            max_chunks: configuration.max_chunks,
-            max_bytes: configuration.max_bytes,
-        }
-    }
-}
-
-impl From<StableConfigurationV2> for Configuration {
-    fn from(stable_configuration: StableConfigurationV2) -> Self {
-        Self {
-            max_batches: stable_configuration.max_batches,
-            max_chunks: stable_configuration.max_chunks,
-            max_bytes: stable_configuration.max_bytes,
-        }
-    }
-}
-
 #[derive(Default)]
 pub struct State {
     assets: HashMap<AssetKey, Asset>,
@@ -356,24 +239,6 @@ pub struct StableStatePermissions {
     manage_permissions: BTreeSet<Principal>,
 }
 
-/// Same as [StableStatePermissions] but serializable with cbor
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct StableStatePermissionsV2 {
-    commit: BTreeSet<Principal>,
-    prepare: BTreeSet<Principal>,
-    manage_permissions: BTreeSet<Principal>,
-}
-
-impl From<StableStatePermissions> for StableStatePermissionsV2 {
-    fn from(stable_state_permissions: StableStatePermissions) -> Self {
-        Self {
-            commit: stable_state_permissions.commit,
-            prepare: stable_state_permissions.prepare,
-            manage_permissions: stable_state_permissions.manage_permissions,
-        }
-    }
-}
-
 #[derive(Clone, Debug, CandidType, Deserialize)]
 pub struct StableState {
     authorized: Vec<Principal>, // ignored if permissions is Some(_)
@@ -382,33 +247,6 @@ pub struct StableState {
 
     next_batch_id: Option<BatchId>,
     configuration: Option<Configuration>,
-}
-
-/// Same as [StableState] but serializable with cbor
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct StableStateV2 {
-    authorized: Vec<Principal>, // ignored if permissions is Some(_)
-    permissions: Option<StableStatePermissionsV2>,
-    stable_assets: HashMap<String, StableAssetV2>,
-
-    next_batch_id: Option<u64>,
-    configuration: Option<StableConfigurationV2>,
-}
-
-impl From<StableState> for StableStateV2 {
-    fn from(stable_state: StableState) -> Self {
-        Self {
-            authorized: stable_state.authorized,
-            permissions: stable_state.permissions.map(Into::into),
-            stable_assets: stable_state
-                .stable_assets
-                .into_iter()
-                .map(|(k, v)| (k, v.into()))
-                .collect(),
-            next_batch_id: stable_state.next_batch_id.map(batch_id_to_u64),
-            configuration: stable_state.configuration.map(Into::into),
-        }
-    }
 }
 
 impl Asset {
@@ -1297,27 +1135,6 @@ impl State {
                 &response.body,
                 None,
             );
-        }
-    }
-}
-
-impl From<State> for StableStateV2 {
-    fn from(state: State) -> Self {
-        let permissions = StableStatePermissionsV2 {
-            commit: state.commit_principals,
-            prepare: state.prepare_principals,
-            manage_permissions: state.manage_permissions_principals,
-        };
-        Self {
-            authorized: vec![],
-            permissions: Some(permissions),
-            stable_assets: state
-                .assets
-                .into_iter()
-                .map(|(k, v)| (k, v.into()))
-                .collect(),
-            next_batch_id: Some(batch_id_to_u64(state.next_batch_id)),
-            configuration: Some(state.configuration.into()),
         }
     }
 }
