+++ src/canisters/frontend/ic-certified-assets/src/tests.rs
@@ -2,7 +2,7 @@ use crate::CreateChunksArg;
 use crate::asset_certification::types::http::{
     CallbackFunc, HttpRequest, HttpResponse, StreamingCallbackToken, StreamingStrategy,
 };
-use crate::state_machine::{BATCH_EXPIRY_NANOS, StableState, State};
+use crate::state_machine::{BATCH_EXPIRY_NANOS, StableStateV2, State};
 use crate::types::{
     AssetProperties, BatchId, BatchOperation, CommitBatchArguments, CommitProposedBatchArguments,
     ComputeEvidenceArguments, CreateAssetArguments, CreateChunkArg, DeleteAssetArguments,
@@ -938,7 +938,7 @@ fn preserves_state_on_stable_roundtrip() {
         ],
     );
 
-    let stable_state: StableState = state.into();
+    let stable_state: StableStateV2 = state.into();
     let state: State = stable_state.into();
 
     let response = certified_http_request(
@@ -1591,7 +1591,7 @@ fn alias_behavior_persists_through_upgrade() {
         SUBDIR_INDEX_BODY
     );
 
-    let stable_state: StableState = state.into();
+    let stable_state: StableStateV2 = state.into();
     let state: State = stable_state.into();
 
     let alias_stays_turned_off = state.http_request(
